Separated heuristic defnition and tightened other logic
